event_project/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via django
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
django==4.2.9
    # via
    #   -r requirements.in
    #   django-extensions
    #   djangorestframework
    #   drf-spectacular
    #   drf-spectacular-sidecar
django-environ==0.11.2
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular==0.27.1
    # via -r requirements.in
drf-spectacular-sidecar==2024.1.1
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.21.1
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
pytz==2023.4
    # via djangorestframework
pyyaml==6.0.1
    # via drf-spectacular
referencing==0.33.0
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.17.1
    # via
    #   jsonschema
    #   referencing
sqlparse==0.4.4
    # via django
uritemplate==4.1.1
    # via drf-spectacular
